Consolidate all logging stuff into one module
